+++ sei-ibc-go/modules/core/exported/client.go
@@ -2,10 +2,9 @@ package exported
 
 import (
 	ics23 "github.com/confio/ics23/go"
-	proto "github.com/gogo/protobuf/proto"
-
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	proto "github.com/gogo/protobuf/proto"
 )
 
 // Status represents the status of a client